+++ StarLevelSystem/Data/LocationResetData.cs
@@ -972,7 +972,7 @@ namespace StarLevelSystem.Data {
                     ResetHours = 72f,
                     Members = new List<string>() {
                         "FortressRuins", "CharredTowerRuins1_dvergr",
-                        "CharredRuins1", "CharredRuins4", "FortressRuins",
+                        "CharredRuins1", "CharredRuins4",
                         "CharredTowerRuins1", "CharredTowerRuins2", "CharredTowerRuins3",
                         "SulfurArch", "CharredRuins2", "CharredRuins3", "VoltureNest"
                     },